feat: add Quotes API for fetching and syncing market data
Add API endpoints for:
- GET /quotes/latest/{symbol} - fetch latest quote from data source
- GET /quotes/history/{symbol} - fetch historical quotes from data source
- POST /quotes/sync/{symbol} - sync quotes from data source to database

Also fix DailyQuote model to use Integer instead of BigInteger for
id column to ensure SQLite compatibility in tests.

+++ src/db/models.py
@@ -116,7 +116,7 @@ class DailyQuote(Base):
     """Daily stock quotes table."""
     __tablename__ = "daily_quotes"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     symbol: Mapped[str] = mapped_column(String(20), nullable=False)
     market: Mapped[Market] = mapped_column(Enum(Market), nullable=False)
     trade_date: Mapped[date] = mapped_column(Date, nullable=False, index=True)
